+++ lib/karafka/web/pro/ui/controllers/topics/base_controller.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+# This code is part of Karafka Pro, a commercial component not licensed under LGPL.
+# See LICENSE for details.
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Controllers
+          # Namespace for all the controllers related to working with topics
+          module Topics
+            # Base controller for all topics related controllers
+            class BaseController < Controllers::BaseController
+            end
+          end
+        end
+      end
+    end
+  end
+end
