--- lib/karafka/web/pro/ui/controllers/topics_controller.rb
@@ -1,122 +1,0 @@
-# frozen_string_literal: true
-
-# This code is part of Karafka Pro, a commercial component not licensed under LGPL.
-# See LICENSE for details.
-
-module Karafka
-  module Web
-    module Pro
-      module Ui
-        module Controllers
-          # Topics management controller
-          # Allows for exploration of settings and replication details
-          class TopicsController < BaseController
-            self.sortable_attributes = %w[
-              name
-              value
-              default?
-              read_only?
-              synonym?
-              sensitive?
-              partition_id
-              leader
-              replica_count
-              in_sync_replica_brokers
-              count
-              share
-              diff
-              low
-              high
-            ].freeze
-
-            # Lists available topics in the cluster
-            def index
-              @topics = Models::Topic.all.sort_by(&:topic_name)
-
-              unless ::Karafka::Web.config.ui.visibility.internal_topics
-                @topics.delete_if { |topic| topic[:topic_name].start_with?('__') }
-              end
-
-              render
-            end
-
-            # Displays requested topic config details
-            #
-            # @param topic_name [String] topic we're interested in
-            def config(topic_name)
-              @topic = Models::Topic.find(topic_name)
-
-              @configs = refine(@topic.configs)
-
-              render
-            end
-
-            # Displays requested topic replication details
-            #
-            # @param topic_name [String] topic we're interested in
-            def replication(topic_name)
-              @topic = Models::Topic.find(topic_name)
-
-              @partitions = refine(@topic[:partitions])
-
-              render
-            end
-
-            # Displays the messages distribution across various partitions
-            #
-            # @param topic_name [String] topic we're interested in
-            #
-            # @note Because computing distribution is fairly expensive, we paginate this. While
-            #   because of that results may not be exact, this allows us to support topics with
-            #   many partitions.
-            def distribution(topic_name)
-              @topic = Models::Topic.find(topic_name)
-
-              @active_partitions, _materialized_page, @limited = Paginators::Partitions.call(
-                @topic.partition_count, @params.current_page
-              )
-
-              @aggregated, distribution = @topic.distribution(@active_partitions)
-
-              @distribution = refine(distribution)
-
-              next_page = @active_partitions.last < @topic.partition_count - 1
-              paginate(@params.current_page, next_page)
-
-              render
-            end
-
-            # Displays high and low offsets for given topic
-            #
-            # @param topic_name [String] topic we're interested in
-            def offsets(topic_name)
-              @topic = Models::Topic.find(topic_name)
-
-              @active_partitions, _materialized_page, @limited = Paginators::Partitions.call(
-                @topic.partition_count, @params.current_page
-              )
-
-              offsets = @active_partitions.map do |partition_id|
-                part_offsets = Admin.read_watermark_offsets(topic_name, partition_id)
-
-                {
-                  partition_id: partition_id,
-                  low: part_offsets.first,
-                  high: part_offsets.last,
-                  diff: part_offsets.last - part_offsets.first
-                }
-              end
-
-              @offsets = refine(offsets)
-
-              next_page = @active_partitions.last < @topic.partition_count - 1
-              paginate(@params.current_page, next_page)
-
-              render
-            end
-          end
-        end
-      end
-    end
-  end
-end
