--- lib/karafka/web/pro/ui/views/topics/_breadcrumbs.erb
@@ -1,48 +1,0 @@
-<%# This code is part of Karafka Pro, a commercial component not licensed under LGPL. %>
-<%# See LICENSE for details. %>
-
-<li>
-  <a href="<%= root_path('topics') %>">
-    Topics
-  </a>
-</li>
-
-<% if @topic %>
-  <li>
-    <a href="<%= root_path('topics', @topic.topic_name, 'config') %>">
-      <%= @topic.topic_name %>
-    </a>
-  </li>
-<% end %>
-
-<% if @configs %>
-  <li>
-    <a href="<%= root_path('topics', @topic.topic_name, 'config') %>">
-      Configuration
-    </a>
-  </li>
-<% end %>
-
-<% if @partitions %>
-  <li>
-    <a href="<%= root_path('topics', @topic.topic_name, 'replication') %>">
-      Replication
-    </a>
-  </li>
-<% end %>
-
-<% if @distribution %>
-  <li>
-    <a href="<%= root_path('topics', @topic.topic_name, 'distribution') %>">
-      Distribution
-    </a>
-  </li>
-<% end %>
-
-<% if @offsets %>
-  <li>
-    <a href="<%= root_path('topics', @topic.topic_name, 'offsets') %>">
-      Offsets
-    </a>
-  </li>
-<% end %>
