--- lib/karafka/web/pro/ui/views/topics/_partition_offsets.erb
@@ -1,13 +1,0 @@
-<%# This code is part of Karafka Pro, a commercial component not licensed under LGPL. %>
-<%# See LICENSE for details. %>
-
-<tr>
-  <td>
-    <a href="<%= explorer_path(@topic[:topic_name], partition_offsets[:partition_id]) %>">
-      <%= partition_offsets[:partition_id] %>
-    </a>
-  </td>
-  <td><%= partition_offsets[:low] %></td>
-  <td><%= partition_offsets[:high] %></td>
-  <td><%= partition_offsets[:diff] %></td>
-</tr>
