+++ lib/karafka/web/pro/ui/views/topics/configs/_breadcrumbs.erb
@@ -1,0 +1,22 @@
+<%# This code is part of Karafka Pro, a commercial component not licensed under LGPL. %>
+<%# See LICENSE for details. %>
+
+<li>
+  <a href="<%= root_path('topics') %>">
+    Topics
+  </a>
+</li>
+
+<li>
+  <a href="<%= root_path('topics', @topic.topic_name, 'config') %>">
+    <%= @topic.topic_name %>
+  </a>
+</li>
+
+<% if @configs %>
+  <li>
+    <a href="<%= root_path('topics', @topic.topic_name, 'config') %>">
+      Configuration
+    </a>
+  </li>
+<% end %>
