+++ lib/karafka/web/pro/ui/views/topics/configs/show.erb
@@ -3,23 +3,24 @@
 
 <% view_title @topic.topic_name %>
 
-<%== partial 'topics/tabs' %>
+<%== partial 'topics/topics/tabs' %>
 
 <div class="data-table-wrapper">
   <table class="data-table">
     <thead>
       <tr>
-        <th><%== sort_link(:partition_id) %></th>
-        <th><%== sort_link(:low) %></th>
-        <th><%== sort_link(:high) %></th>
-        <th><%== sort_link('Count', :diff) %></th>
+        <th><%== sort_link(:name) %></th>
+        <th><%== sort_link(:value) %></th>
+        <th><%== sort_link(:default?) %></th>
+        <th><%== sort_link(:sensitive?) %></th>
+        <th><%== sort_link(:read_only?) %></th>
       </tr>
     </thead>
     <tbody>
       <%==
         each_partial(
-          @offsets,
-          'topics/partition_offsets'
+          @configs,
+          'cluster/config'
         )
       %>
     </tbody>
