+++ lib/karafka/web/pro/ui/views/topics/distributions/_empty_partitions.erb
