+++ lib/karafka/web/pro/ui/views/topics/distributions/_partition.erb
