+++ lib/karafka/web/pro/ui/views/topics/offsets/show.erb
@@ -3,24 +3,23 @@
 
 <% view_title @topic.topic_name %>
 
-<%== partial 'topics/tabs' %>
+<%== partial 'topics/topics/tabs' %>
 
 <div class="data-table-wrapper">
   <table class="data-table">
     <thead>
       <tr>
-        <th><%== sort_link(:name) %></th>
-        <th><%== sort_link(:value) %></th>
-        <th><%== sort_link(:default?) %></th>
-        <th><%== sort_link(:sensitive?) %></th>
-        <th><%== sort_link(:read_only?) %></th>
+        <th><%== sort_link(:partition_id) %></th>
+        <th><%== sort_link(:low) %></th>
+        <th><%== sort_link(:high) %></th>
+        <th><%== sort_link('Count', :diff) %></th>
       </tr>
     </thead>
     <tbody>
       <%==
         each_partial(
-          @configs,
-          'cluster/config'
+          @offsets,
+          'topics/offsets/partition'
         )
       %>
     </tbody>
