+++ lib/karafka/web/pro/ui/views/topics/replications/_partition.erb
