+++ lib/karafka/web/pro/ui/views/topics/topics/_breadcrumbs.erb
@@ -1,0 +1,8 @@
+<%# This code is part of Karafka Pro, a commercial component not licensed under LGPL. %>
+<%# See LICENSE for details. %>
+
+<li>
+  <a href="<%= root_path('topics') %>">
+    Topics
+  </a>
+</li>
