+++ lib/karafka/web/pro/ui/views/topics/topics/_tabs.erb
