+++ lib/karafka/web/pro/ui/views/topics/topics/_topic.erb
