+++ lib/karafka/web/pro/ui/views/topics/topics/index.erb
@@ -10,7 +10,7 @@
     <%==
       each_partial(
         @topics,
-        'topics/topic'
+        'topics/topics/topic'
       )
     %>
   </div>
