+++ spec/lib/karafka/web/pro/ui/controllers/topics/topics_controller_spec.rb
@@ -1,0 +1,66 @@
+# frozen_string_literal: true
+
+# This code is part of Karafka Pro, a commercial component not licensed under LGPL.
+# See LICENSE for details.
+
+RSpec.describe_current do
+  subject(:app) { Karafka::Web::Pro::Ui::App }
+
+  let(:internal_topic) { "__#{generate_topic_name}" }
+
+  describe '#index' do
+    before do
+      create_topic(topic_name: internal_topic)
+      get 'topics'
+    end
+
+    it do
+      expect(response).to be_ok
+      expect(body).to include(breadcrumbs)
+      expect(body).not_to include(pagination)
+      expect(body).not_to include(support_message)
+      expect(body).to include(topics_config.consumers.states)
+      expect(body).to include(topics_config.consumers.metrics)
+      expect(body).to include(topics_config.consumers.reports)
+      expect(body).to include(topics_config.errors)
+      expect(body).not_to include(internal_topic)
+    end
+
+    context 'when there are no topics' do
+      before do
+        allow(::Karafka::Web::Ui::Models::Topic).to receive(:all).and_return([])
+        get 'topics'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include('There are no available topics in the current cluster')
+      end
+    end
+
+    context 'when internal topics should be displayed' do
+      before do
+        allow(::Karafka::Web.config.ui.visibility)
+          .to receive(:internal_topics)
+          .and_return(true)
+
+        get 'topics'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include(topics_config.consumers.states)
+        expect(body).to include(topics_config.consumers.metrics)
+        expect(body).to include(topics_config.consumers.reports)
+        expect(body).to include(topics_config.errors)
+        expect(body).to include(internal_topic)
+      end
+    end
+  end
+end
